Jueun 파일 생성
