+++ main.py
@@ -1,97 +1,1 @@
-import time
-import random
-
-# player_number는 인원수, userinex는 PlayerList에서 사용자의 인덱스번호, PlayerList는 게임하는 사람의 리스트
-def gameJueun(player_number, userindex, playerList):
-    #지하철 게임
-    Subway = [['연천', '전곡', '청산', '소요산', '동두천', '보산', '동두천중앙', '지행', '덕정', '덕계', '양주', '녹양', '가능', '의정부',
-               '회룡', '망월사', '도봉산', '도봉', '방학', '창동', '녹천', '월계', '광운대', '석계', '신이문', '외대앞', '회기', '청량리',
-               '제기동', '신설동', '동묘앞', '동대문', '종로5가', '종로3가', '종각', '시청', '서울', '남영', '용산', '노량진', '대방', '신길',
-               '영등포', '신도림', '구로', '가산디지털단지', '독산', '금천구청', '광명', '석수', '관악', '안양', '명학', '금정', '군포', '당정',
-               '의왕', '성균관대', '화서', '수원', '세류', '병점', '서동탄', '세마', '오산대', '오산', '진위', '송탄', '서정리', '평탱지제',
-               '평택', '성환', '직산', '두정', '천안', '봉명', '쌍용', '아산', '탕정', '배방', '온양온천', '신창', '구일', '개봉', '오류동',
-               '온수', '역곡', '소사', '부천', '중동', '송내', '부개', '부평', '백운', '동암', '간석', '주안', '도화', '제물포', '도원',
-               '동인천', '인천'],
-                ['강남', '역삼', '선릉', '삼성', '종합운동장', '잠실새내', '잠실', '잠실나루', '강변', '구의', '건대입구', '성수', '뚝섬',
-                '한양대', '왕십리', '상왕십리', '신당', '동대문역사문화공원', '을지로4가', '을지로3가', '을지로입구', '시청', '충정로', '아현',
-                '이대', '신촌', '홍대입구', '합정', '당산', '영등포구청', '문래', '신도림', '대림', '구로디지털단지', '신대방', '신림', '봉천',
-                '서울대입구', '낙성대', '사당', '방배', '서초', '교대', '용답', '신답', '용두', '신설동', '도림천', '양천구청', '신정네거리',
-                '까치산'], 
-                ['대화', '주엽', '정발산', '마두', '백석', '대곡', '화정', '원당', '원흥', '삼송', '지축', '구파밭', '연신내', '불광', '녹번',
-                 '홍제', '무악재', '독립문', '경북궁', '안국', '종로3가', '을지로3가', '충무로', '동대입구', '약수', '금호', '옥수', '입구정',
-                 '신사', '잠원', '고속터미널', '교대', '남부터미널', '양재', '매봉', '도곡', '대치', '학여울', '대청', '일원', '수서', '가락시장',
-                 '경찰병원', '오금'],
-                ['진접', '오남', '별내별가람', '불암산', '상계', '노원', '창동', '쌍문', '수유', '미아', '미아사거리', '길음', '성신여대입구',
-                  '한성대입구', '혜화', '동대문', '동대문역사문화공원', '충무로', '명동', '회현', '서울', '숙대입구', '삼각지', '신용산', '이촌',
-                  '동작', '총신대입구', '이수', '사당', '남태령', '선바위', '경마공원', '대공원', '과천', '정부과천청사', '인덕원', '평촌', '범계',
-                  '금정', '산본', '수리산', '대야미', '반월', '상록수', '한대앞', '중앙', '고잔', '초지', '안산', '신길온천', '정왕', '오이도'],
-                ['방화', '개화산', '김포공항', '송정', '마곡', '발산', '우장산', '화곡', '까치산', '신정', '목동', '오목교', '양평', '영등포구청',
-                 '영등포시장', '신길', '여의도', '여의나루', '마포', '공덕', '애오개', '충정로', '서대문', '광화문', '종로3가', '을지로4가',
-                 '동대문역사문화공원', '청구', '신금호', '행당', '왕십리', '마장', '답십리', '장한평', '군자', '아차산', '광나루', '천호', '강동',
-                 '둔춘동', '올림픽공원', '방이', '오금', '개롱', '거여', '마천', '길동', '굽은다리', '명일', '고덕', '상일동', '강일', '미사',
-                 '하남풍산', '하남시청', '하남검단산'],
-                ['응암', '역촌', '불광', '독바위', '연신내', '구산', '새절', '증산', '디지털미디어시티', '월드컵경기장', '마포구청', '망원', '합정',
-                 '상수', '광흥창', '대흥', '공덕', '효창공원앞', '삼각지', '녹사평', '이태원', '한강진', '버티고개', '약수', '청구', '신당', '동묘앞',
-                 '창신', '보문', '안암', '고려대', '월곡', '상월곡', '돌곶이', '석계', '태릉입구', '화랑대', '봉화산', '신내'],
-                 ['장암', '도봉산', '수락산', '마들', '노원', '중계', '하계', '공릉','태릉입구', '먹골', '중화', '사가정', '용마산', '중곡', '군자',
-                  '어린이대공원', '건대입구', '자양', '청담', '강남구청', '학동', '논현', '반포', '고속터미널', '내방', '이수', '남성', '숭실대입구',
-                  '상도', '장승배기', '신대방삼거리', '보라매', '신풍', '대림', '남구로', '가산디지털단지', '철산', '광명사거리', '천왕', '온수', '까치울',
-                  '부천종합운동장', '춘의', '신중동', '부천시청', '상동', '삼산체육관', '굴포천', '부평구청', '산곡', '석남'],
-                  ['별내', '다산', '동구릉', '구리', '장자호수공원', '암사역사공원', '암사', '천호', '강동구청', '몽촌토성', '잠실', '석촌',
-                   '송파', '가락시장', '문정', '장지', '복정', '남위례', '산성', '남한산성입구', '단대오거리', '신흥', '수진', '모란'],
-                   ['개화', '김포공항', '공항시장', '신방화', '마곡나루', '양천향교', '가양', '증미', '등촌', '염창', '신목동', '선유도', '당산',
-                    '국회의사당', '여의도', '샛강', '노량진', '노들', '흑석', '동작', '구반포', '신반포', '고속터미널', '사평', '신논현', '언주',
-                    '선정릉', '삼성중앙', '봉은사', '종합운동장', '삼전','석촌고분', '석촌', '송파나루', '한성백제', '올림픽공원',
-                    '둔촌오륜', '중앙보훈병원']]
-    check = [0] * 100
-
-    print("지하철~지하철~지하철~지하철~🚆🚆🚆")
-
-    try:
-        n = int(input("몇~호선~? 몇~호선~? 몇~호선~? 몇~호선~?(1, 2, 3, 4, 5, 6, 7, 8, 9중에서 선택)"))
-        if not n in [1, 2, 3, 4, 5, 6, 7, 8, 9]:
-            raise Exception
-    except:
-        #입력을 제대로 받지 않았을 때
-        print("병신샷~ 병신샷~")
-        print(f"아 누가누가 술을 마셔 {playerList[0]}이(가) 술을 마셔 원~~~샷")
-        return
-    else :
-        print(f"{n}호선~{n}호션~{n}호선~{n}호선")
-
-    int(n)
-    print("게임 시작~ (주의: 3초 제한)")
-    while True:
-        for i in range(player_number):
-            # i가 user인지 확인하기
-            if i == userindex :
-                start_time = time.time()
-                ans = input(f"{playerList[i]} : ")
-                end_time = time.time()
-                #3초 제한
-                if (end_time - start_time > 3) :
-                    print("3초 지남ㅋ")
-                    print(f"아 누가누가 술을 마셔 {playerList[i]}이(가) 술을 마셔 원~~~샷")
-                    return playerList[i]
-                # Subway에 없는 것을 입력했을 때
-                elif not ans in Subway[n-1] :
-                    print(f"아 누가누가 술을 마셔 {playerList[i]}이(가) 술을 마셔 원~~~샷")
-                    return playerList[i]
-                # 이미 누군가 했던 지하철 역을 입력했을 때
-                elif check[Subway[n-1].index(ans)] == 1:
-                    print(f"아 누가누가 술을 마셔 {playerList[i]}이(가) 술을 마셔 원~~~샷")
-                    return playerList[i]
-                else :
-                    # 한번 나온 지하철 역은 check하기
-                    check[Subway[n-1].index(ans)] = 1
-            else :
-                #랜덤으로 체크되지 않은 지하철 역이 나올 때까지
-                while True:
-                    r = random.randint(0, len(Subway[n-1]))
-                    if check[r] == 0:
-                        break
-                print(f"{playerList[i]} : {Subway[n-1][r]}")
-                check[r] = 1
-
-
-gameJueun(3, 0, ['주은', '지은', '동수'])
+import Jueun
